abstracted searchbar component
